Move post execution back to TestExecutionState
